api/requirements.txt
asgiref==3.8.1
Django==4.2.14
djangorestframework==3.15.2
flake8==7.1.0
mccabe==0.7.0
psycopg2-binary==2.9.3
pycodestyle==2.12.0
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
